# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: blindkiss API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [upcoming, past, all]
            default: all
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events, admin]
      summary: Create an event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    put:
      operationId: updateEvent
      tags: [events, admin]
      summary: Update event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events, admin]
      summary: Delete event (admin)
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /content/{key}:
    parameters:
      - in: path
        name: key
        required: true
        schema:
          type: string
    put:
      operationId: updateContentBlock
      tags: [content, admin]
      summary: Update content block (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentBlockInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentBlock"
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    Event:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        venue: { type: string }
        city: { type: string }
        address: { type: string }
        mapsUrl: { type: string, nullable: true }
        ticketUrl: { type: string, nullable: true }
        price: { type: string, nullable: true }
        eventDate: { type: string, format: date-time }
        posterUrl: { type: string, nullable: true }
        description: { type: string, nullable: true }
        forcePast: { type: boolean }
        isPast: { type: boolean }
      required: [id, title, venue, city, address, eventDate, forcePast, isPast]
    EventInput:
      type: object
      properties:
        title: { type: string }
        venue: { type: string }
        city: { type: string }
        address: { type: string }
        mapsUrl: { type: string, nullable: true }
        ticketUrl: { type: string, nullable: true }
        price: { type: string, nullable: true }
        eventDate: { type: string, format: date-time }
        posterUrl: { type: string, nullable: true }
        description: { type: string, nullable: true }
        forcePast: { type: boolean }
      required: [title, venue, city, address, eventDate]
...
    ContentBlock:
      type: object
      properties:
        id: { type: integer }
        key: { type: string }
        title: { type: string, nullable: true }
        body: { type: string, nullable: true }
        imageUrl: { type: string, nullable: true }
      required: [id, key]
    ContentBlockInput:
      type: object
      properties:
        title: { type: string, nullable: true }
        body: { type: string, nullable: true }
        imageUrl: { type: string, nullable: true }
...