# lib/api-zod/src/generated/types/contentBlock.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * blindkiss API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContentBlock {
  id: number;
  key: string;
  title?: string | null;
  body?: string | null;
  imageUrl?: string | null;
}