# lib/api-zod/src/generated/types/contentBlockInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * blindkiss API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContentBlockInput {
  title?: string | null;
  body?: string | null;
  imageUrl?: string | null;
}